requirements.txt
acres
aiofiles
altair
amqp
annotated-types
anyio
asgiref
attrs
beautifulsoup4
billiard
blinker
cachetools
celery
certifi
charset-normalizer
ci-info
click
click-didyoumean
click-plugins
click-repl
configobj
configparser
contourpy
cycler
Deprecated
distlib
Django
django-debug-toolbar
dnspython
docopt
easyocr
email_validator
et_xmlfile
etelemetry
fastapi
fastapi-cli
ffmpy
filelock
fitz
Flask
flask-cors
fonttools
fpdf
fsspec
ghostscript
gitdb
GitPython
gradio
gradio_client
h11
httpcore
httplib2
httptools
httpx
huggingface-hub
idna
imageio
importlib_metadata
importlib_resources
iniconfig
isodate
itsdangerous
Jinja2
Js2Py
jsonschema
jsonschema-specifications
kiwisolver
kombu
lazy_loader
looseversion
lxml
markdown-it-py
MarkupSafe
matplotlib
mdurl
mpmath
networkx
nibabel
ninja
nipype
numpy
opencv-contrib-python
opencv-python
opencv-python-headless
openpyxl
orjson
packaging
pandas
pathlib
pdf-compressor
pdf2image
pikepdf
pillow
pipwin
platformdirs
pluggy
poppler-utils
prompt_toolkit
protobuf
prov
psycopg2-binary
puremagic
pyarrow
pyclipper
pydantic
pydantic_core
pydeck
pydot
pydub
Pygments
pyjsparser
PyMuPDF
PyMuPDFb
pyparsing
PyPDF2
PyPrind
pySmartDL
pytesseract
pytest
python-bidi
python-dateutil
python-dotenv
python-multipart
pytz
pyxnat
PyYAML
rdflib
redis
referencing
requests
rich
rpds-py
ruff
scikit-image
scipy
semantic-version
setuptools
shapely
shellingham
simplejson
six
smmap
sniffio
soupsieve
sqlparse
starlette
streamlit
sympy
tabulate
tenacity
tifffile
tk
toml
tomlkit
toolz
torch
torchvision
tornado
tqdm
traits
typer
typing_extensions
tzdata
tzlocal
ujson
urllib3
uvicorn
uvloop
validators
vine
virtualenv
virtualenvwrapper-win
watchfiles
wcwidth
websockets
Werkzeug
wheel
wrapt
zipp
